# frontend/src/api/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Ground Truth Curator",
		"version": "1.0.6"
	},
	"paths": {
...
		"/v1/schemas/{name}": {
			"get": {
				"tags": ["schemas"],
				"summary": "Get Schema",
				"description": "Return a specific component schema by name from the OpenAPI document.",
				"operationId": "get_schema_v1_schemas__name__get",
				"parameters": [
					{
						"name": "name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Name"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": true,
									"title": "Response Get Schema V1 Schemas  Name  Get"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
		"/v1/tags": {
			"get": {
				"tags": ["tags"],
				"summary": "Get Tags",
				"operationId": "get_tags_v1_tags_get",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TagListResponse"
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["tags"],
				"summary": "Post Tags",
				"operationId": "post_tags_v1_tags_post",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddTagsRequest"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TagListResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["tags"],
				"summary": "Delete Tags",
				"operationId": "delete_tags_v1_tags_delete",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RemoveTagsRequest"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TagListResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"AddTagsRequest": {
				"properties": {
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Tags"
					}
				},
				"type": "object",
				"required": ["tags"],
				"title": "AddTagsRequest"
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
...
			"RemoveTagsRequest": {
				"properties": {
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Tags"
					}
				},
				"type": "object",
				"required": ["tags"],
				"title": "RemoveTagsRequest"
			},
...
			"TagListResponse": {
				"properties": {
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Tags"
					},
					"computedTags": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Computedtags"
					}
				},
				"type": "object",
				"required": ["tags"],
				"title": "TagListResponse"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					}
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			}
		}
	}
}